- Removed hard-coded root password
-Removed the "null" user (i.e. the first "user" in the getAllUsers) - Esgf-security inserts the table column names as the first user.

- Fixed some page styles

+++ src/java/main/org/esgf/adminui/ManageUsersController.java
@@ -295,7 +295,9 @@ public class ManageUsersController {
         LOG.debug("userName->"+userName);
         
         String userId = uoi.getUserIdFromUserName(userName);
-        uoi.deleteUser(userName);
+        if(userName.equalsIgnoreCase("rootAdmin")) {
+            uoi.deleteUser(userName);
+        }
         
         LOG.debug("------End ManageUsersController deleteUser------");
     }
